Expose 1433 port via docker-compose
Read your really helpful post on https://samwalpole.com/how-to-run-aspnet-core-and-sql-server-from-docker
and hit a problem that if i wanted to connect to the db via Rider or DataGrip, then the TCP port was blocked. Having added port mapping, got a successful connection. Feel free to refuse the PR if it feels unnecessary for your demo